PR Reviews: change SizeOf to own module
