+++ src/Pact/Types/Continuation.hs
@@ -42,6 +42,7 @@ import Data.Aeson (ToJSON(..), FromJSON(..))
 import Pact.Types.ChainId (ChainId)
 import Pact.Types.PactValue
 import Pact.Types.Pretty
+import Pact.Types.SizeOf
 import Pact.Types.Term
 import Pact.Types.Util (lensyToJSON, lensyParseJSON)
 
@@ -59,12 +60,13 @@ data Provenance = Provenance
 instance Pretty Provenance where
   pretty (Provenance c h) = "(chain = \"" <> pretty c <> "\" hash=\"" <> pretty h <> "\")"
 
-instance NFData Provenance
-instance ToJSON Provenance where toJSON = lensyToJSON 2
-instance FromJSON Provenance where parseJSON = lensyParseJSON 2
 instance SizeOf Provenance where
   sizeOf (Provenance chainId modHash) =
     (constructorCost 2) + (sizeOf chainId) + (sizeOf modHash)
+
+instance NFData Provenance
+instance ToJSON Provenance where toJSON = lensyToJSON 2
+instance FromJSON Provenance where parseJSON = lensyParseJSON 2
 
 -- | The type of a set of yielded values of a pact step.
 --
@@ -78,6 +80,7 @@ data Yield = Yield
 instance NFData Yield
 instance ToJSON Yield where toJSON = lensyToJSON 2
 instance FromJSON Yield where parseJSON = lensyParseJSON 2
+
 instance SizeOf Yield where
   sizeOf (Yield dataYield prov) =
     (constructorCost 2) + (sizeOf dataYield) + (sizeOf prov)
