+++ src/Pact/Types/Hash.hs
@@ -33,6 +33,7 @@ import Pact.Types.Pretty
 import Pact.Types.Util
 import Data.Aeson
 import GHC.Generics
+import Pact.Types.SizeOf
 import Data.Text (Text)
 import Data.Text.Encoding (decodeUtf8)
 import Data.Reflection
@@ -55,7 +56,7 @@ import Crypto.Hash.Blake2Native
 -- Within Pact these are blake2b_256 but unvalidated as such,
 -- so other hash values are kosher (such as an ETH sha256, etc).
 newtype Hash = Hash { unHash :: ByteString }
-  deriving (Eq, Ord, Generic, Hashable, Serialize)
+  deriving (Eq, Ord, Generic, Hashable, Serialize,SizeOf)
 
 instance Show Hash where
   show (Hash h) = show $ encodeBase64UrlUnpadded h
